feat: v0.0.3 — 功能扩展、全面审计优化、项目整理
新功能:
- Renewal/Pre-Renewal 双公式系统（防御/命中/ASPD/状态点）
- Rhai 公式引擎（db/formulas/，脚本优先+硬编码回退）
- Web HTTP API（axum，/api/v1/status/players/party）
- Fluent 国际化（locales/zh-CN/en-US）
- 组队经验/物品分配（均分/按等级/轮询）
- Mob 技能系统（db/mob_skill_db.yml + AI 释放）
- 4th Job Trait 属性 + AP 系统 + DB migration v6
- 环境变量配置覆盖（DEVIRUCHI_* 前缀）
- DB import 覆盖层（db/import/）
- 日/夜循环、PIN 码、删除冷却、Hash 校验
- DeviAgent CLI 集成（deviruchi agent 交互模式）
- Agent API 配置化（network.agent_port）

安全与质量:
- SQL 注入防护、Agent API 认证、脚本循环保护
- 溢出防护（saturating 算术、safe_u32/u16 转换）
- 生产代码 unwrap 全部替换为安全模式
- 69 处 log→tracing 迁移
- Clippy 0 警告（46 个全部修复）
- 1,393 测试全部通过

项目整理:
- 合并 assets/ → resources/，构建脚本 → scripts/
- 删除 releases/ 跟踪（~111MB 二进制）
- 移除未使用依赖（once_cell, log, tokio-tungstenite）
- 版本号 0.0.2 → 0.0.3

+++ src/game/map/map_server/player.rs
@@ -212,7 +212,7 @@ impl MapServer {
         // 调用 SkillHandler 执行技能逻辑
         let result = self.skill_handler.use_skill(
             Arc::new(player.clone()),
-            skill_pkt.skill_id as u16,
+            skill_pkt.skill_id,
             1, // 默认技能等级 1
             skill_pkt.target_id,
             &self.map_state,
